Moved a few classes to better fitting folders
