+++ CsCore/PlainNetClassLib/src/Plugins/CsCore/com/csutil/async/TaskHelper.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 25c5b7759e3b2ab409fc59aa8c7aecf4
+guid: 3b250d26b93e8154c8a22220461c76b1
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
