+++ CsCore/PlainNetClassLib/src/Plugins/CsCore/com/csutil/datastructures/StateMachine.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 082cb27e3994f0845a1117b6864d61b9
+guid: 07b4871a5a5962549be24235a3cf62d8
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
