+++ CsCore/PlainNetClassLib/src/Plugins/CsCore/com/csutil/datastructures/keyvaluestore/IKeyValueStoreExtensions.cs
@@ -46,4 +46,5 @@ namespace com.csutil.keyvaluestore {
         }
 
     }
+
 }
