+++ CsCore/PlainNetClassLib/src/Plugins/CsCore/com/csutil/datastructures/keyvaluestore/LiteDbKeyValueStore.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3ca8f5db49f814f41bb5f52094df0b50
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
