+++ CsCore/PlainNetClassLib/src/Plugins/CsCore/com/csutil/datastructures/keyvaluestore/RetryKeyValueStore.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 20bab74208f06a443a2377eb0b9af7b7
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
