+++ CsCore/PlainNetClassLib/src/Plugins/CsCore/com/csutil/logging/StopwatchV2.cs
@@ -1,6 +1,5 @@
 using System;
 using System.Diagnostics;
-using com.csutil.io;
 
 namespace com.csutil {
     public class StopwatchV2 : Stopwatch {
