+++ CsCore/PlainNetClassLib/src/Plugins/CsCore/com/csutil/model/HasId.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 36044cb8e3c658342800d3f2651cb2e6
+guid: bc239d53cbc120e4ab5625dc2f176035
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
